Plot read statistics from the CapSet object
Source: R/AllGenerics.R, R/plotting_functions.R

Plot read statistics from the CapSet object

plotReadStats(CSobject, plotType = "dodge", plotValue = "numbers",
  outFile = NULL)

# S4 method for CapSet
plotReadStats(CSobject, plotType = "dodge",
  plotValue = "numbers", outFile = NULL)

Arguments

CSobject
The CapSet object

plotType
character. The type of plot to make. Choose from "stack" or "dodge" for either a stacked barchart, or a bar chart with "dodged" positions (analogous to ggplot)

plotValue
character. What values to plot. Choose from "numbers" or "proportions". If "proportions" is selected, the proportion of reads w.r.t total demultiplexed reads per sample would be plotted

outFile
character. Output file name. (filename extention would be used to determine type). If outfile not specified, the plot would be retured on the screen

Value

A ggplot object, or a file. Plot showing the number/proportion of reads in each category, per sample

Examples

# load a previously saved CapSet object
cs <- exampleCSobject()

#> Checking de-multiplexed R1 reads

#> Checking de-multiplexed R2 reads

#> Checking mapped file

#> Checking de-duplicated file

plotReadStats(cs, plotType = "dodge", plotValue = "numbers", outFile = "test_numbers.pdf")

#> Plotting following information : samples demult_reads num_mapped num_filtered num_intss

#> 

#> Saving 10.1 x 5.6 in image